// frontend/src/lib/uniswap/generated/tradeApi.ts
export interface components {
    schemas: {
// ...
        CreateSwapRequest: {
            quote: components["schemas"]["ClassicQuote"] | components["schemas"]["WrapUnwrapQuote"] | components["schemas"]["BridgeQuote"];
            /** @description The signed permit. */
            signature?: string;
            /**
             * @deprecated
             * @description Use `refreshGasPrice` instead.
             * @default false
             */
            includeGasInfo: boolean;
            /**
             * @description If true, the gas price will be re-fetched from the network.
             * @default false
             */
            refreshGasPrice: boolean;
            /**
             * @description If true, the transaction will be simulated. If the simulation results on an onchain error, endpoint will return an error.
             * @default false
             */
            simulateTransaction: boolean;
            permitData?: components["schemas"]["Permit"];
            safetyMode?: components["schemas"]["SwapSafetyMode"];
            deadline?: components["schemas"]["deadline"];
            urgency?: components["schemas"]["Urgency"];
        };
// ...
        CreateSwapResponse: {
            requestId: components["schemas"]["RequestId"];
            swap: components["schemas"]["TransactionRequest"];
            gasFee?: components["schemas"]["gasFee"];
        };
// ...
        QuoteResponse: {
            requestId: components["schemas"]["RequestId"];
            quote: components["schemas"]["Quote"];
            routing: components["schemas"]["Routing"];
            permitTransaction?: components["schemas"]["TransactionRequest"];
            permitData: components["schemas"]["NullablePermit"];
            permitGasFee?: components["schemas"]["gasFee"];
        };
        LimitOrderQuoteResponse: {
            requestId: components["schemas"]["RequestId"];
            quote: components["schemas"]["DutchQuote"];
            /** @enum {string} */
            routing: "LIMIT_ORDER";
            permitData: components["schemas"]["NullablePermit"];
        };
        QuoteRequest: {
            type: components["schemas"]["TradeType"];
            amount: components["schemas"]["tokenAmount"];
            tokenInChainId: components["schemas"]["ChainId"];
            tokenOutChainId: components["schemas"]["ChainId"];
            tokenIn: components["schemas"]["inputToken"];
            tokenOut: components["schemas"]["outputToken"];
            generatePermitAsTransaction?: components["schemas"]["generatePermitAsTransaction"];
            swapper: components["schemas"]["senderWalletAddress"];
            slippageTolerance?: components["schemas"]["slippageTolerance"];
            autoSlippage?: components["schemas"]["AutoSlippage"];
            routingPreference?: components["schemas"]["RoutingPreference"];
            protocols?: components["schemas"]["Protocols"];
            hooksOptions?: components["schemas"]["HooksOptions"];
            spreadOptimization?: components["schemas"]["SpreadOptimization"];
            urgency?: components["schemas"]["Urgency"];
            permitAmount?: components["schemas"]["PermitAmount"];
        };
// ...
        OrderResponse: {
            requestId: components["schemas"]["RequestId"];
            orderId: components["schemas"]["orderId"];
            orderStatus: components["schemas"]["OrderStatus"];
        };
        OrderRequest: {
            /** @description The signed permit. */
            signature: string;
            quote: components["schemas"]["DutchQuoteV2"] | components["schemas"]["DutchQuoteV3"] | components["schemas"]["PriorityQuote"];
            routing?: components["schemas"]["Routing"];
        };
// ...
        Routing: "DUTCH_LIMIT" | "CLASSIC" | "DUTCH_V2" | "DUTCH_V3" | "BRIDGE" | "LIMIT_ORDER" | "PRIORITY" | "WRAP" | "UNWRAP" | "CHAINED";
// ...
        ApprovalRequest: {
            walletAddress: components["schemas"]["senderWalletAddress"];
            token: components["schemas"]["inputToken"];
            amount: components["schemas"]["tokenAmount"];
            chainId: components["schemas"]["ChainId"];
            urgency?: components["schemas"]["Urgency"];
            includeGasInfo?: components["schemas"]["includeGasInfo"];
            tokenOut?: components["schemas"]["outputToken"];
            tokenOutChainId?: components["schemas"]["ChainId"];
        };
        ApprovalResponse: {
            requestId: components["schemas"]["RequestId"];
            approval: components["schemas"]["TransactionRequest"];
            cancel: components["schemas"]["TransactionRequest"];
            gasFee?: components["schemas"]["gasFee"];
            cancelGasFee?: components["schemas"]["gasFee"];
        };
// ...
        TransactionRequest: {
            to: components["schemas"]["receiverWalletAddress"];
            from: components["schemas"]["senderWalletAddress"];
            /** @description The calldata for the transaction. */
            data: string;
            /** @description The quantity of ETH tokens approved for spending by the transaction, denominated in wei. Note that by default Uniswap Labs sets this to the maximum approvable spend. */
            value: string;
            gasLimit?: components["schemas"]["gasLimit"];
            chainId: components["schemas"]["ChainId"];
            maxFeePerGas?: components["schemas"]["maxFeePerGas"];
            maxPriorityFeePerGas?: components["schemas"]["maxPriorityFeePerGas"];
            gasPrice?: components["schemas"]["gasPrice"];
        };
// ...
        WalletCheckDelegationRequestBody: {
            /** @description Array of wallet addresses to check delegation status for. */
            walletAddresses?: components["schemas"]["Address"][];
            /** @description Array of chain IDs to check delegation status for. */
            chainIds: components["schemas"]["ChainId"][];
        };
// ...
        WalletCheckDelegationResponseBody: {
            requestId: components["schemas"]["RequestId"];
            /** @description Map of wallet addresses to chain IDs to delegation details. */
            delegationDetails: {
                [key: string]: components["schemas"]["ChainDelegationMap"];
            };
        };
// ...
    };
// ...
}